chore: Add is secure tests.
